moves PosterBookmark to molecules
