+++ src/components/molecules/PosterBookmark/index.ts
